chore: reorganize src/helpers/ into focused files under src/utils/
- Rename directory src/helpers/ → src/utils/
- Split helpers.ts into four focused files:
  - utils/vscode.ts — safeGlobalStateGet/Update, isRedHatBuild
  - utils/process.ts — CLI execution helpers, CommandOutput interface
  - utils/path.ts — arePathsEqual, findFolderOfPomXml, resolvePaths
  - utils/version.ts — normalizeVersionForSemver
- Move all other helper files unchanged into src/utils/
- Update all import sites across extension, commands, executors,
  services, webview, and test files
