+++ src/test/services/CamelLauncherDownloader.test.ts
@@ -2,7 +2,7 @@ import { assert } from 'chai';
 import * as path from 'path';
 import * as fs from 'fs';
 import { CamelLauncherDownloader, LauncherNotFoundError } from '../../services/CamelLauncherDownloader';
-import { isRedHatBuild } from '../../helpers/helpers';
+import { isRedHatBuild } from '../../utils/vscode';
 
 suite('CamelLauncherDownloader Tests', () => {
 	let downloader: CamelLauncherDownloader;
