+++ src/test/webview/suggestionProviders.test.ts
@@ -16,7 +16,7 @@
  */
 import path from 'path';
 import { expect } from 'chai';
-import { filterSuggestionsByWord, getSuggestions } from '../../helpers/SuggestionRegistry';
+import { filterSuggestionsByWord, getSuggestions } from '../../utils/SuggestionRegistry';
 import { Suggestion } from '@kaoto/kaoto';
 
 suite('Channel API', () => {
