+++ src/utils/ApplicationPropertiesFinder.ts
