+++ src/utils/DockerErrorDetector.ts
