+++ src/utils/SuggestionRegistry.ts
